Make sure folder or files exist in build

+++ build/build.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
 
-rm bin/bootstrap
-rm bin/cleeper.zip
+if [ ! -d "bin" ]; then
+  mkdir bin
+fi
+
+if [ -f bin/bootstrap ]; then
+    rm bin/bootstrap
+fi
+
+if [ -f bin/cleeper.zip ]; then
+    rm bin/cleeper.zip
+fi
+
 cd ../src/
 go mod tidy
 GOARCH=amd64 GOOS=linux CGO_ENABLED=0 go build -ldflags="-s -w" -o ../build/bin/bootstrap
